pyproject.toml
[tool.poetry]
name = "vembrane"
version = "0.3.2"
description = "Filter VCF/BCF files with Python expressions."
authors = ["Till Hartmann"]
readme = "README.md"
homepage = "https://github.com/vembrane/vembrane"
repository = "https://github.com/vembrane/vembrane"

[tool.poetry.scripts]
vembrane = 'vembrane.cli:main'

[tool.poetry.dependencies]
python = "^3.7"
pysam = "^0.16"
pyyaml = "^5.3"
importlib_metadata = {version = "^1.7.0", python = "<3.8"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyyaml = "^5.3"
pre-commit = "^1.10"
black = { version = "^19.10b0", python = "^3.7" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
